Dialogue Implementation Changes
Created a new map, game mode, and several new functions and assets to enable the creation of dialogue trees that allow for the player to provide responses, and the NPC's dialogue to branch based on player input. System isn't very automated right now, but anyone could assemble an NPC conversation out of the components I've created, using the two example Dialogue Trees as a guide.