DamageTiles_V2
Continued to work on the damage tiles task.
I made a temporary setup of checking if the tiles along the journey have a pass through value that is not 0 and then deal damage via the add damage function.
Then, I made a temp setup to check the last index of the tiles array to see if it has a stood value that is not 0 and then pass the value to take damage